+++ python/subprojects/boosting/mlrl/boosting/boosting_learners.py
@@ -76,7 +76,7 @@ class Boomer(RuleLearner, ClassifierMixin, RegressorMixin, MultiOutputMixin):
                  random_state: int = 1,
                  feature_format: str = SparsePolicy.AUTO.value,
                  label_format: str = SparsePolicy.AUTO.value,
-                 predicted_label_format: str = SparsePolicy.AUTO.value,
+                 prediction_format: str = SparsePolicy.AUTO.value,
                  statistic_format: Optional[str] = None,
                  default_rule: Optional[str] = None,
                  rule_induction: Optional[str] = None,
@@ -183,7 +183,7 @@ class Boomer(RuleLearner, ClassifierMixin, RegressorMixin, MultiOutputMixin):
                                                 parallel or not. Must be 'true' or 'false'. For additional options refer
                                                 to the documentation
         """
-        super().__init__(random_state, feature_format, label_format, predicted_label_format)
+        super().__init__(random_state, feature_format, label_format, prediction_format)
         self.statistic_format = statistic_format
         self.default_rule = default_rule
         self.rule_induction = rule_induction